FEAT: Create Template Makefile
This commit introduces a template Makefile for the implementation of an
easy, consistent compiling solution that can support multiple included
source files and subdirectories. When calling the Makefile using `make`,
the user can input the intended output PDF filename, which would be a
combination of the basename for their source TeX file and the `.pdf`
file extension. For example, `make example-ko.pdf` would compile
`example-ko.tex` (and its included files) into "example-ko.pdf".

As of now, the default option when calling `make` is 'example-ko.pdf',
the Korean example document in the `example/` directory.
While the preferable way of calling `make` would be to specify
the document explicitly, this can be improved by manually changing
the target document filename to the name of the main file.
For instance, if the main `.tex` file is to be named "main.tex",
the `all: ` target would be changed to `main.pdf`, to ensure that
`main.pdf` is compiled when `make` is called without an explicit target.

Additional variables supported in this Makefile are `TEXDIST`, as the
engine to compile the document, and `TEXFLAGS`, as additional flags fed
to the engine during compilation.
The `TEXDIST` variable can be `lua` or `xe` to call `lualatex` and
`xelatex` respectively.
The `TEXFLAGS` variable can be a set of flags directly fed to the
engine, such as `--shell-escape` for `minted` package users.

The Makefile also contains a `clean` target, which is a ".PHONY" target;
these are targets not bound to any specific output file.
The current design of `make clean` is to remove any auxiliary filetypes
that are not tracked by Git, such as `.aux` auxiliary files,
`.log` logfiles and `.toc` table-of-contents logs.

Further improvements can be made to this Makefile; one good idea may be
to implement a compiling option without refreshing the bibliography
database. This is mentioned in the Makefile as the `nobib` option.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>