FEAT: Add Default Styles for Abstract Titles
Centered titles and a 3-centimeter horizontal skip for Korean
is specified in the ECE standard template.
To ensure that centering and horizontal skip do not affect the ToC,
this commit introduces constants `\l_snu_ece_bsc_thesis_abstractstyle_*`
for each supported language.
These "style" constants are used to typeset chapter titles conformant to
the provided standards, while the original "name" constants are passed
to the ToC line.

To further support abstract environments without an explicit language
option, an extra string comparison condition is implemented.
This condition checks if the `\abstractname` has been modified -
if not, the predefined "name" constants are passed to the ToC by default
to mimic behaviour for environments with an explicit language option
provided.
If the `\abstractname` has been modified, the modified `\abstractname`
is passed as-is to the ToC.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>