FEAT: Remove Vertical Skips and Center Chapters
In further accordance with the ECE standards, chapter titles must be
offset by 3 centimeters from the top of the page,
with its contents centered and typeset in a single line.
The current base class `report` does not support such changes without
modifying the underlying macros, effectively "hard-coding" some aspects
that require changes.

To circumvent this problem, a new package `titlesec` is introduced as a
prerequisite for this class;
macros `\titlespacing` and `\titleformat` are used to modify vertical
spacing and formatting, respectively.

This removes the necessity for abstract titles to be centered
explicitly, as their `\chapter*{...}`s are also centered by default
with the new `\titleformat` in effect.

Note that the `\titlespacing` macro has no effect on `\chapter`s unless
the `\titleformat` macro is also used.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>